+++ integration/backends/secrets_test.go
@@ -42,9 +42,13 @@ func TestSecretsE2E(t *testing.T) {
 	err = migrate.Migrate(db, mysql.Mem)
 	require.NoError(t, err)
 
-	topic, err := pubsub.OpenTopic(ctx, fmt.Sprintf("%s://secrets-test", mempubsub.Scheme))
+	jobTopic, err := pubsub.OpenTopic(ctx, fmt.Sprintf("%s://secrets-test-jobs", mempubsub.Scheme))
 	require.NoError(t, err)
-	defer topic.Shutdown(ctx)
+	defer jobTopic.Shutdown(ctx)
+
+	checkTopic, err := pubsub.OpenTopic(ctx, fmt.Sprintf("%s://secrets-test-checks", mempubsub.Scheme))
+	require.NoError(t, err)
+	defer checkTopic.Shutdown(ctx)
 
 	ur := mysql.NewUserRepository(db)
 	tr := mysql.NewTeamRepository(db)
@@ -59,7 +63,7 @@ func TestSecretsE2E(t *testing.T) {
 	suow := unitwork.NewStartUnitOfWork(db, mysql.Mem)
 
 	jwtSecret := []byte("test-secret")
-	svc := pikoci.New(ctx, topic, ur, tr, ppr, jr, rr, rt, br, rur, str, tgr, suow, jwtSecret, logger)
+	svc := pikoci.New(ctx, jobTopic, checkTopic, ur, tr, ppr, jr, rr, rt, br, rur, str, tgr, suow, jwtSecret, logger)
 	svc.StartScheduler(ctx)
 
 	// Migration already creates admin user and "main" team.
@@ -71,15 +75,19 @@ func TestSecretsE2E(t *testing.T) {
 	}, true)
 
 	// Start worker
-	subscription, err := pubsub.OpenSubscription(ctx, fmt.Sprintf("%s://secrets-test", mempubsub.Scheme))
+	jobSub, err := pubsub.OpenSubscription(ctx, fmt.Sprintf("%s://secrets-test-jobs", mempubsub.Scheme))
 	require.NoError(t, err)
-	defer subscription.Shutdown(ctx)
+	defer jobSub.Shutdown(ctx)
+
+	checkSub, err := pubsub.OpenSubscription(ctx, fmt.Sprintf("%s://secrets-test-checks", mempubsub.Scheme))
+	require.NoError(t, err)
+	defer checkSub.Shutdown(ctx)
 
 	var wg sync.WaitGroup
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		w := worker.New(svc, topic, subscription, logger.With("component", "worker"))
+		w := worker.New(svc, jobTopic, jobSub, checkSub, logger.With("component", "worker"))
 		w.Run(ctx)
 	}()