+++ integration/main_test.go
@@ -47,11 +47,17 @@ func runTests(m *testing.M) int {
 		panic(err)
 	}
 
-	topic, err := pubsub.OpenTopic(ctx, getTopicURL(mempubsub.Scheme))
+	jobTopic, err := pubsub.OpenTopic(ctx, getTopicURL(mempubsub.Scheme, "pikoci-jobs"))
 	if err != nil {
-		panic(fmt.Errorf("failed to open: %v", err).Error())
+		panic(fmt.Errorf("failed to open job topic: %v", err).Error())
 	}
-	defer topic.Shutdown(ctx)
+	defer jobTopic.Shutdown(ctx)
+
+	checkTopic, err := pubsub.OpenTopic(ctx, getTopicURL(mempubsub.Scheme, "pikoci-checks"))
+	if err != nil {
+		panic(fmt.Errorf("failed to open check topic: %v", err).Error())
+	}
+	defer checkTopic.Shutdown(ctx)
 
 	ur := mysql.NewUserRepository(db)
 	tr := mysql.NewTeamRepository(db)
@@ -64,7 +70,7 @@ func runTests(m *testing.M) int {
 	str := mysql.NewSecretTypeRepository(db)
 	tgr := mysql.NewTriggerRepository(db)
 	suow := unitwork.NewStartUnitOfWork(db, mysql.Mem)
-	var svc = pikoci.New(ctx, topic, ur, tr, ppr, jr, rr, rt, br, rur, str, tgr, suow, jwtSecret, logger)
+	var svc = pikoci.New(ctx, jobTopic, checkTopic, ur, tr, ppr, jr, rr, rt, br, rur, str, tgr, suow, jwtSecret, logger)
 	svc.StartScheduler(ctx)
 	var handler = tshttp.Handler(svc, jwtSecret, logger.With("component", "HTTP"), db, mysql.Mem)
 	server := httptest.NewServer(handler)
@@ -75,14 +81,14 @@ func runTests(m *testing.M) int {
 	_, _ = svc.CreateUser(ctx, user.User{FullName: "pepito", Username: "pepito", Password: "$2a$14$rwQk8Qvc2rij7qhFO4P1W.OiSF6AkgVU1RCrLaY2wawJcpkPEKwbm"}, isHash)
 	_, _ = svc.CreateUser(ctx, user.User{FullName: "grillo", Username: "grillo", Password: "$2a$14$SvWir17.jlXxiZfe0pJuDedznetc/HWKv43YPsQQNo6MJiuypS2q6"}, isHash)
 	go func() {
-		runWorker(ctx, mempubsub.Scheme, topic, svc, 1, "DEBUG")
+		runWorker(ctx, mempubsub.Scheme, jobTopic, svc, 1, "DEBUG")
 	}()
 
 	return m.Run()
 }
 
-func getTopicURL(s string) string {
-	u := fmt.Sprintf("%s://pikoci", s)
+func getTopicURL(s, name string) string {
+	u := fmt.Sprintf("%s://%s", s, name)
 	switch s {
 	case natspubsub.Scheme:
 		u += "?natsv2"
@@ -90,16 +96,16 @@ func getTopicURL(s string) string {
 	return u
 }
 
-func getSubscriptionURL(s string) string {
-	u := fmt.Sprintf("%s://pikoci", s)
+func getSubscriptionURL(s, name string) string {
+	u := fmt.Sprintf("%s://%s", s, name)
 	switch s {
 	case natspubsub.Scheme:
-		u += "?queue=pikoci&natsv2"
+		u += fmt.Sprintf("?queue=%s&natsv2", name)
 	}
 	return u
 }
 
-func runWorker(ctx context.Context, sy string, t queue.Topic, s pikoci.Service, c int, llvl string) error {
+func runWorker(ctx context.Context, sy string, jobTopic queue.Topic, s pikoci.Service, c int, llvl string) error {
 	var lvl slog.Level
 	switch llvl {
 	case "DEBUG":
@@ -112,19 +118,25 @@ func runWorker(ctx context.Context, sy string, t queue.Topic, s pikoci.Service, 
 		lvl = slog.LevelInfo
 	}
 	logger := slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: lvl})).With("service", "worker")
-	// Create a subscription connected to that topic.
-	subscription, err := pubsub.OpenSubscription(ctx, getSubscriptionURL(sy))
+
+	jobSub, err := pubsub.OpenSubscription(ctx, getSubscriptionURL(sy, "pikoci-jobs"))
 	if err != nil {
-		return fmt.Errorf("failed to OpenSubscription: %w", err)
+		return fmt.Errorf("failed to open job subscription: %w", err)
 	}
-	defer subscription.Shutdown(ctx)
+	defer jobSub.Shutdown(ctx)
+
+	checkSub, err := pubsub.OpenSubscription(ctx, getSubscriptionURL(sy, "pikoci-checks"))
+	if err != nil {
+		return fmt.Errorf("failed to open check subscription: %w", err)
+	}
+	defer checkSub.Shutdown(ctx)
 
 	var wg sync.WaitGroup
 	for i := range c {
 		wg.Add(1)
 		nlogger := logger.With("num", i+1)
 		nlogger.Info(fmt.Sprintf("Starting Worker %d", i+1))
-		w := worker.New(s, t, subscription, nlogger)
+		w := worker.New(s, jobTopic, jobSub, checkSub, nlogger)
 
 		go func() {
 			err = w.Run(ctx)